+++ chispa/formatting/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from chispa.formatting.formats import RESET, Color, Format, FormattingConfig, Style
-from chispa.formatting.terminal_string_formatter import format_string
+from chispa.formatting.format_string import format_string
+from chispa.formatting.formats import RESET, Color, Format, Style
+from chispa.formatting.formatting_config import FormattingConfig
 
 __all__ = ("Style", "Color", "FormattingConfig", "Format", "format_string", "RESET")
