+++ chispa/formatting/formats.py
@@ -2,12 +2,16 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from enum import Enum
-from typing import ClassVar
 
 RESET = "\033[0m"
 
 
 class Color(str, Enum):
+    """
+    Enum for terminal colors.
+    Each color is represented by its corresponding ANSI escape code.
+    """
+
     BLACK = "\033[30m"
     RED = "\033[31m"
     GREEN = "\033[32m"
@@ -27,6 +31,11 @@ class Color(str, Enum):
 
 
 class Style(str, Enum):
+    """
+    Enum for text styles.
+    Each style is represented by its corresponding ANSI escape code.
+    """
+
     BOLD = "\033[1m"
     UNDERLINE = "\033[4m"
     BLINK = "\033[5m"
@@ -36,13 +45,32 @@ class Style(str, Enum):
 
 @dataclass
 class Format:
+    """
+    Data class to represent text formatting with color and style.
+
+    Attributes:
+        color (Color | None): The color for the text.
+        style (list[Style] | None): A list of styles for the text.
+    """
+
     color: Color | None = None
     style: list[Style] | None = None
 
     @classmethod
     def from_dict(cls, format_dict: dict) -> Format:
+        """
+        Create a Format instance from a dictionary.
+
+        Args:
+            format_dict (dict): A dictionary with keys 'color' and/or 'style'.
+        """
         if not isinstance(format_dict, dict):
             raise ValueError("Input must be a dictionary")
+
+        valid_keys = {"color", "style"}
+        invalid_keys = set(format_dict) - valid_keys
+        if invalid_keys:
+            raise ValueError(f"Invalid keys in format dictionary: {invalid_keys}. Valid keys are {valid_keys}")
 
         color = cls._get_color_enum(format_dict.get("color"))
         style = format_dict.get("style")
@@ -54,6 +82,34 @@ class Format:
             styles = None
 
         return cls(color=color, style=styles)
+
+    @classmethod
+    def from_list(cls, values: list[str]) -> Format:
+        """
+        Create a Format instance from a list of strings.
+
+        Args:
+            values (list[str]): A list of strings representing colors and styles.
+        """
+        if not all(isinstance(value, str) for value in values):
+            raise ValueError("All elements in the list must be strings")
+
+        color = None
+        styles = []
+        valid_colors = [c.name.lower() for c in Color]
+        valid_styles = [s.name.lower() for s in Style]
+
+        for value in values:
+            if value in valid_colors:
+                color = Color[value.upper()]
+            elif value in valid_styles:
+                styles.append(Style[value.upper()])
+            else:
+                raise ValueError(
+                    f"Invalid value: {value}. Valid values are colors: {valid_colors} and styles: {valid_styles}"
+                )
+
+        return cls(color=color, style=styles if styles else None)
 
     @staticmethod
     def _get_color_enum(color: Color | str | None) -> Color | None:
@@ -78,50 +134,3 @@ class Format:
                 valid_styles = [f.name.lower() for f in Style]
                 raise ValueError(f"Invalid style name: {style}. Valid style names are {valid_styles}")
         return None
-
-
-class FormattingConfig:
-    """
-    Class to manage and parse formatting configurations.
-    """
-
-    VALID_KEYS: ClassVar = {"color", "style"}
-
-    def __init__(
-        self,
-        mismatched_rows: Format | dict = Format(Color.RED),
-        matched_rows: Format | dict = Format(Color.BLUE),
-        mismatched_cells: Format | dict = Format(Color.RED, [Style.UNDERLINE]),
-        matched_cells: Format | dict = Format(Color.BLUE),
-    ):
-        """
-        Initializes the FormattingConfig with given or default formatting.
-
-        Each of the arguments can be provided as a `Format` object or a dictionary with the following keys:
-        - 'color': A string representing a color name, which should be one of the valid colors:
-            ['black', 'red', 'green', 'yellow', 'blue', 'purple', 'cyan', 'light_gray',
-            'dark_gray', 'light_red', 'light_green', 'light_yellow', 'light_blue',
-            'light_purple', 'light_cyan', 'white'].
-        - 'style': A string or list of strings representing styles, which should be one of the valid styles:
-            ['bold', 'underline', 'blink', 'invert', 'hide'].
-
-        Args:
-            mismatched_rows (Format | dict): Format or dictionary for mismatched rows.
-            matched_rows (Format | dict): Format or dictionary for matched rows.
-            mismatched_cells (Format | dict): Format or dictionary for mismatched cells.
-            matched_cells (Format | dict): Format or dictionary for matched cells.
-
-        Raises:
-            ValueError: If the dictionary contains invalid keys or values.
-        """
-        self.mismatched_rows: Format = self._parse_format(mismatched_rows)
-        self.matched_rows: Format = self._parse_format(matched_rows)
-        self.mismatched_cells: Format = self._parse_format(mismatched_cells)
-        self.matched_cells: Format = self._parse_format(matched_cells)
-
-    def _parse_format(self, format: Format | dict) -> Format:
-        if isinstance(format, Format):
-            return format
-        elif isinstance(format, dict):
-            return Format.from_dict(format)
-        raise ValueError("Invalid format type. Must be Format or dict.")
