+++ tests/formatting/test_formats.py
@@ -2,86 +2,9 @@ from __future__ import annotations
 
 import re
 
-from chispa.formatting.formats import Color, Format, FormattingConfig, Style
+import pytest
 
-
-def test_default_mismatched_rows():
-    config = FormattingConfig()
-    assert config.mismatched_rows.color == Color.RED
-    assert config.mismatched_rows.style is None
-
-
-def test_default_matched_rows():
-    config = FormattingConfig()
-    assert config.matched_rows.color == Color.BLUE
-    assert config.matched_rows.style is None
-
-
-def test_default_mismatched_cells():
-    config = FormattingConfig()
-    assert config.mismatched_cells.color == Color.RED
-    assert config.mismatched_cells.style == [Style.UNDERLINE]
-
-
-def test_default_matched_cells():
-    config = FormattingConfig()
-    assert config.matched_cells.color == Color.BLUE
-    assert config.matched_cells.style is None
-
-
-def test_custom_mismatched_rows():
-    config = FormattingConfig(mismatched_rows={"color": "green", "style": ["bold", "underline"]})
-    assert config.mismatched_rows.color == Color.GREEN
-    assert config.mismatched_rows.style == [Style.BOLD, Style.UNDERLINE]
-
-
-def test_custom_matched_rows():
-    config = FormattingConfig(matched_rows={"color": "yellow"})
-    assert config.matched_rows.color == Color.YELLOW
-    assert config.matched_rows.style is None
-
-
-def test_custom_mismatched_cells():
-    config = FormattingConfig(mismatched_cells={"color": "purple", "style": ["blink"]})
-    assert config.mismatched_cells.color == Color.PURPLE
-    assert config.mismatched_cells.style == [Style.BLINK]
-
-
-def test_custom_matched_cells():
-    config = FormattingConfig(matched_cells={"color": "cyan", "style": ["invert", "hide"]})
-    assert config.matched_cells.color == Color.CYAN
-    assert config.matched_cells.style == [Style.INVERT, Style.HIDE]
-
-
-def test_invalid_color():
-    try:
-        FormattingConfig(mismatched_rows={"color": "invalid_color"})
-    except ValueError as e:
-        assert (
-            str(e)
-            == "Invalid color name: invalid_color. Valid color names are ['black', 'red', 'green', 'yellow', 'blue', 'purple', 'cyan', 'light_gray', 'dark_gray', 'light_red', 'light_green', 'light_yellow', 'light_blue', 'light_purple', 'light_cyan', 'white']"
-        )
-
-
-def test_invalid_style():
-    try:
-        FormattingConfig(mismatched_rows={"style": ["invalid_style"]})
-    except ValueError as e:
-        assert (
-            str(e)
-            == "Invalid style name: invalid_style. Valid style names are ['bold', 'underline', 'blink', 'invert', 'hide']"
-        )
-
-
-def test_invalid_key():
-    try:
-        FormattingConfig(mismatched_rows={"invalid_key": "value"})
-    except ValueError as e:
-        error_message = str(e)
-        assert re.match(
-            r"Invalid keys in format dictionary: \{'invalid_key'\}. Valid keys are \{('color', 'style'|'style', 'color')\}",
-            error_message,
-        )
+from chispa.formatting import Color, Format, Style
 
 
 def test_format_from_dict_valid():
@@ -93,24 +16,23 @@ def test_format_from_dict_valid():
 
 def test_format_from_dict_invalid_color():
     format_dict = {"color": "invalid_color", "style": ["bold"]}
-    try:
+    with pytest.raises(ValueError) as exc_info:
         Format.from_dict(format_dict)
-    except ValueError as e:
-        assert (
-            str(e)
-            == "Invalid color name: invalid_color. Valid color names are ['black', 'red', 'green', 'yellow', 'blue', 'purple', 'cyan', 'light_gray', 'dark_gray', 'light_red', 'light_green', 'light_yellow', 'light_blue', 'light_purple', 'light_cyan', 'white']"
-        )
+    assert str(exc_info.value) == (
+        "Invalid color name: invalid_color. Valid color names are "
+        "['black', 'red', 'green', 'yellow', 'blue', 'purple', 'cyan', 'light_gray', "
+        "'dark_gray', 'light_red', 'light_green', 'light_yellow', 'light_blue', 'light_purple', "
+        "'light_cyan', 'white']"
+    )
 
 
 def test_format_from_dict_invalid_style():
     format_dict = {"color": "blue", "style": ["invalid_style"]}
-    try:
+    with pytest.raises(ValueError) as exc_info:
         Format.from_dict(format_dict)
-    except ValueError as e:
-        assert (
-            str(e)
-            == "Invalid style name: invalid_style. Valid style names are ['bold', 'underline', 'blink', 'invert', 'hide']"
-        )
+    assert str(exc_info.value) == (
+        "Invalid style name: invalid_style. Valid style names are " "['bold', 'underline', 'blink', 'invert', 'hide']"
+    )
 
 
 def test_format_from_dict_invalid_key():
@@ -123,3 +45,55 @@ def test_format_from_dict_invalid_key():
             r"Invalid keys in format dictionary: \{'invalid_key'\}. Valid keys are \{('color', 'style'|'style', 'color')\}",
             error_message,
         )
+
+
+def test_format_from_list_valid():
+    values = ["blue", "bold", "underline"]
+    format_instance = Format.from_list(values)
+    assert format_instance.color == Color.BLUE
+    assert format_instance.style == [Style.BOLD, Style.UNDERLINE]
+
+
+def test_format_from_list_invalid_color():
+    values = ["invalid_color", "bold"]
+    with pytest.raises(ValueError) as exc_info:
+        Format.from_list(values)
+    assert str(exc_info.value) == (
+        "Invalid value: invalid_color. Valid values are colors: "
+        "['black', 'red', 'green', 'yellow', 'blue', 'purple', 'cyan', 'light_gray', "
+        "'dark_gray', 'light_red', 'light_green', 'light_yellow', 'light_blue', 'light_purple', "
+        "'light_cyan', 'white'] and styles: ['bold', 'underline', 'blink', 'invert', 'hide']"
+    )
+
+
+def test_format_from_list_invalid_style():
+    values = ["blue", "invalid_style"]
+    with pytest.raises(ValueError) as exc_info:
+        Format.from_list(values)
+    assert str(exc_info.value) == (
+        "Invalid value: invalid_style. Valid values are colors: "
+        "['black', 'red', 'green', 'yellow', 'blue', 'purple', 'cyan', 'light_gray', "
+        "'dark_gray', 'light_red', 'light_green', 'light_yellow', 'light_blue', 'light_purple', "
+        "'light_cyan', 'white'] and styles: ['bold', 'underline', 'blink', 'invert', 'hide']"
+    )
+
+
+def test_format_from_list_non_string_elements():
+    values = ["blue", 123]
+    with pytest.raises(ValueError) as exc_info:
+        Format.from_list(values)
+    assert str(exc_info.value) == "All elements in the list must be strings"
+
+
+def test_format_from_dict_empty():
+    format_dict = {}
+    format_instance = Format.from_dict(format_dict)
+    assert format_instance.color is None
+    assert format_instance.style is None
+
+
+def test_format_from_list_empty():
+    values = []
+    format_instance = Format.from_list(values)
+    assert format_instance.color is None
+    assert format_instance.style is None
